Session_ID change

+++ app.py
@@ -41,9 +41,11 @@ app.secret_key = os.urandom(24)
 # create_table()
 
 
-def get_phrase():
+def get_phrase(session_id):
     # if 'user' in session:
-    session_id = session['user']
+    #     session_ID = session['user']
+    #     return session_ID
+    # # session_ID = session_ID
     # session_id = session_ID
     phraseget = Sessions.query.filter_by(session_id=session_id).first()
     phrase = phraseget.phrase
@@ -68,9 +70,11 @@ def get_phrase():
     return phrase
 
 
-def get_word_guess():
+def get_word_guess(session_id):
     # if 'user' in session:
-    session_id = session['user']
+    #     session_ID = session['user']
+    #     return session_ID
+    # # session_ID = session_ID
     # session_id = session_ID
     word_guessget = Sessions.query.filter_by(session_id=session_id).first()
     word_guess = word_guessget.word_guess
@@ -97,10 +101,11 @@ def get_word_guess():
     return word_guess
 
 
-def get_alpl():
+def get_alpl(session_id):
     # if 'user' in session:
-    session_id = session['user']
-
+    #     session_ID = session['user']
+    #     return session_ID
+    # # session_ID = session_ID
     # session_id = session_ID
     alplget = Sessions.query.filter_by(session_id=session_id).first()
     alpl = alplget.alpl
@@ -128,9 +133,11 @@ def get_alpl():
     return alpl
 
 
-def get_turns():
+def get_turns(session_id):
     # if 'user' in session:
-    session_id = session['user']
+    #     session_ID = session['user']
+    #     return session_ID
+    # # session_ID = session_ID
     # session_id = session_ID
     turnsget = Sessions.query.filter_by(session_id=session_id).first()
     turns = turnsget.turns
@@ -150,9 +157,11 @@ def get_turns():
     return turns
 
 
-def get_politician():
+def get_politician(session_id):
     # if 'user' in session:
-    session_id = session['user']
+    #     session_ID = session['user']
+    #     return session_ID
+    # # session_ID = session_ID
     # session_id = session_ID
     politicianget = Sessions.query.filter_by(session_id=session_id).first()
     politician = politicianget.politician
@@ -181,8 +190,11 @@ def Hangman():
         # session['user'] = 'Anthony'
         session['user'] = random.getrandbits(50)
 
-        if 'user' in session:
-            session_ID = session['user']
+        # if 'user' in session:
+        session_ID = session['user']
+        print(session_ID)
+            # return session_ID
+
 
         if 'user' not in session:
             session['user'] = random.getrandbits(50)
@@ -190,8 +202,12 @@ def Hangman():
         # conn = sqlite3.connect('politicalhangman.db')
         # c = conn.cursor()
 
-        def session_id_enter():
+        def session_id_enter(session_ID):
 
+            # if 'user' in session:
+            #     session_ID = session['user']
+            #     return session_ID
+            # # session_ID = session_ID
             session_id = session_ID
             datestamp = str(datetime.now())
             id_enter = Sessions(session_id=session_id, datestamp=datestamp, politician=None, politician_id=None,
@@ -206,7 +222,9 @@ def Hangman():
             #             (session_id, datestamp))
             # conn.commit()
 
-        session_id_enter()
+
+        print(session_ID)
+        session_id_enter(session_ID)
 
         # Game resets on returning to the front page
         # global word_guess
@@ -238,6 +256,7 @@ def Politician_choice():
 
         # if 'user' in session:
 
+
         return render_template('Politician_choice.html')
 
     elif request.method == 'POST':
@@ -250,6 +269,9 @@ def Politician_choice():
         from pol_ids import t_dict
 
         politician_id = t_dict[politician]
+
+        session_ID = session['user']
+
 
         # conn = sqlite3.connect('politicalhangman.db')
         # c = conn.cursor()
@@ -363,7 +385,7 @@ def Politician_choice():
         # conn = sqlite3.connect('politicalhangman.db')
         # c = conn.cursor()
 
-        def session_info_enter(politician, politician_id, word_guess, phrase, alpl, turns):
+        def session_info_enter(politician, politician_id, word_guess, phrase, alpl, turns, session_ID):
 
             # remaining_letters =''.join(remaining_letters)
             # if 'user' in session:
@@ -373,12 +395,17 @@ def Politician_choice():
 
             word_guess = ''.join(word_guess)
             # if 'user' in session:
-            session_ID = session['user']
-            # c.execute("UPDATE game_db set politician=?, politician_id=?, word_guess=?, phrase=?, alpl=?, turns=? WHERE session_id=?",
-            #           [politician, politician_id, word_guess, phrase, alpl, turns, session_ID])
-            # conn.commit()
-            # c.close()
-            # conn.close()
+            # session_ID = session['user']
+            # # c.execute("UPDATE game_db set politician=?, politician_id=?, word_guess=?, phrase=?, alpl=?, turns=? WHERE session_id=?",
+            # #           [politician, politician_id, word_guess, phrase, alpl, turns, session_ID])
+            # # conn.commit()
+            # # c.close()
+            # # conn.close()
+            # session_id = session_ID
+            # if 'user' in session:
+            #     session_ID = session['user']
+            #     return session_ID
+            # session_ID = session_ID
             session_id = session_ID
             sessioninfo = Sessions.query.filter_by(session_id=session_id).first()
             sessioninfo.politician = politician
@@ -390,7 +417,7 @@ def Politician_choice():
             db.session.commit()
             # print('session_info_enter')
 
-        session_info_enter(politician, politician_id, word_guess, phrase, alpl, turns)
+        session_info_enter(politician, politician_id, word_guess, phrase, alpl, turns, session_ID)
 
         return redirect(url_for('Game'))
 
@@ -401,11 +428,17 @@ def About():
 
         # conn = sqlite3.connect('politicalhangman.db')
         # c = conn.cursor()
+        print('get')
 
         def delete_db_session():
 
             # if 'user' in session:
             #     session_ID = session['user']
+            # session_id = session_ID
+            if 'user' in session:
+                session_ID = session['user']
+                return session_ID
+            # session_ID = session_ID
             session_id = session_ID
             deletesession = Sessions.query.filter_by(session_id=session_id).first()
             db.session.delete(deletesession)
@@ -431,11 +464,19 @@ def About():
 def Game():
     if request.method == 'GET':
 
-        phrase = get_phrase()
-        word_guess = get_word_guess()
-        alpl = get_alpl()
-        turns = get_turns()
-        politician = get_politician()
+        # if 'user' in session:
+        #     session_ID = session['user']
+        #     return session_ID
+        #     # session_ID = session_ID
+        # session_id = session_ID
+        session_id = session['user']
+
+        phrase = get_phrase(session_id)
+        word_guess = get_word_guess(session_id)
+        alpl = get_alpl(session_id)
+        turns = get_turns(session_id)
+        politician = get_politician(session_id)
+
 
         # print('Test123')
         # print(word_guess)
@@ -484,21 +525,27 @@ def Game():
         # conn = sqlite3.connect('politicalhangman.db')
         # c = conn.cursor()
 
-        def session_info_enter(turns, guess_phrase):
+        def session_info_enter(turns, guess_phrase, session_id):
             # if 'user' in session:
-            session_ID = session['user']
-            # c.execute("UPDATE game_db set turns=?, guess_phrase=? WHERE session_id=?",
-            #           [turns, guess_phrase, session_ID])
-            # conn.commit()
-            # c.close()
-            # conn.close()
-            session_id = session_ID
+            # session_ID = session['user']
+            # # c.execute("UPDATE game_db set turns=?, guess_phrase=? WHERE session_id=?",
+            # #           [turns, guess_phrase, session_ID])
+            # # conn.commit()
+            # # c.close()
+            # # conn.close()
+            # session_id = session_ID
+            # if 'user' in session:
+            #     session_ID = session['user']
+            #     return session_ID
+            # session_ID = session_ID
+            # session_id = session_ID
+            print('session_info_enter')
             sessioninfo = Sessions.query.filter_by(session_id=session_id).first()
             sessioninfo.turns = turns
             sessioninfo.guess_phrase = guess_phrase
             db.session.commit()
 
-        session_info_enter(turns, guess_phrase)
+        session_info_enter(turns, guess_phrase, session_id)
 
         guess_head = 'The latest tweet from %r' % politician
 
@@ -514,11 +561,13 @@ def Game():
 
         # Get guess+phrase, convert to list,
 
-        phrase = get_phrase()
-        word_guess = get_word_guess()
-        alpl = get_alpl()
-        turns = get_turns()
-        politician = get_politician()
+        session_id = session['user']
+
+        phrase = get_phrase(session_id)
+        word_guess = get_word_guess(session_id)
+        alpl = get_alpl(session_id)
+        turns = get_turns(session_id)
+        politician = get_politician(session_id)
 
         # GEt all variables from game_db return variables
 
@@ -587,6 +636,11 @@ def Game():
 
                 # if 'user' in session:
                 #     session_ID = session['user']
+                # session_id = session_ID
+                if 'user' in session:
+                    session_ID = session['user']
+                    return session_ID
+                # session_ID = session_ID
                 session_id = session_ID
                 deletesession = Sessions.query.filter_by(session_id=session_id).first()
                 db.session.delete(deletesession)
@@ -614,6 +668,11 @@ def Game():
             def delete_db_session():
                 # if 'user' in session:
                 #     session_ID = session['user']
+                # session_id = session_ID
+                if 'user' in session:
+                    session_ID = session['user']
+                    return session_ID
+                # session_ID = session_ID
                 session_id = session_ID
                 deletesession = Sessions.query.filter_by(session_id=session_id).first()
                 db.session.delete(deletesession)
@@ -632,7 +691,7 @@ def Game():
         # conn = sqlite3.connect('politicalhangman.db')
         # c = conn.cursor()
 
-        def session_info_enter(turns, guess_phrase, alpl, phrase, word_guess):
+        def session_info_enter(turns, guess_phrase, alpl, phrase, word_guess, session_id):
             alpl = "".join(alpl)
             word_guess = ''.join(word_guess)
             # if 'user' in session:
@@ -644,8 +703,13 @@ def Game():
             # conn.close()
 
             # if 'user' in session:
-            session_ID = session['user']
-            session_id = session_ID
+            # session_ID = session['user']
+            # session_id = session_ID
+            # if 'user' in session:
+            #     session_ID = session['user']
+            #     return session_ID
+            # # session_ID = session_ID
+            # session_id = session_ID
             session_update = Sessions.query.filter_by(session_id=session_id).first()
             session_update.turns = turns
             session_update.guess_phrase = guess_phrase
@@ -654,7 +718,7 @@ def Game():
             session_update.word_guess = word_guess
             db.session.commit()
 
-        session_info_enter(turns, guess_phrase, alpl, phrase, word_guess)
+        session_info_enter(turns, guess_phrase, alpl, phrase, word_guess, session_id)
 
         # print(guess_phrase)
